pkg/cpuset: fix file descriptor leak
Worker processes were dying because of too many open files. Inspection
showed the leak was is opening the shield cpuset tasks file.

$ sudo ls -l /proc/28634/fd | awk '{print $11}' | sort | uniq -c
      1
      1 anon_inode:[eventpoll]
      3 /dev/pts/3
      2 pipe:[7794087]
      1 pipe:[7812340]
      1 pipe:[7812341]
      1 socket:[7791497]
      1 socket:[7792434]
    759 /sys/fs/cgroup/cpuset/sys/tasks
      1 /tmp/contbench694245590/sandbox/bench028699768/bench.out

This suggested an error somewhere in pkg/cpuset, and indeed there was a
bug in `writefile` where it would not close the file in the event that
the write syscall failed.

This commit defers the close call, ensuring the file descriptor is
always closed.